requirements.txt
pylast==7.0.2
python-dotenv==1.1.0
ttkbootstrap==1.20.1
tksvg==0.7.4